+++ src/entities/company/index.ts
@@ -1,8 +1,10 @@
-export { companyAdvantages, companyOverview } from './model/company-overview'
-export type { CompanyAdvantage } from './model/company-overview'
+export { companyApproachPrinciples, companyApproachSection } from './model/company-approach'
+export type { CompanyApproachPrinciple } from './model/company-approach'
 export {
   companyContactSection,
   companyPhoneNumbers,
   companySocialLinks,
 } from './model/company-contacts'
 export type { CompanySocialLink } from './model/company-contacts'
+export { companyProofBlocks, companyProofSection } from './model/company-proof'
+export type { CompanyProofBlock } from './model/company-proof'