+++ src/shared/config/navigation/navigation-items.ts
@@ -4,11 +4,11 @@ export type NavigationItem = {
 }
 
 export const navigationItems: NavigationItem[] = [
-  { href: '/#home', label: 'Главная' },
-  { href: '/#assortment', label: 'Ассортимент и услуги' },
-  { href: '/#about', label: 'О нас' },
-  { href: '/#projects', label: 'Наши проекты' },
-  { href: '/#contacts', label: 'Контакты' },
+  { href: '/#home', label: 'Home' },
+  { href: '/#solutions', label: 'Solutions' },
+  { href: '/#why-us', label: 'Why Us' },
+  { href: '/#projects', label: 'Projects' },
+  { href: '/#contact', label: 'Contact' },
 ]
 
-export const contactNavigationItems = navigationItems.filter((item) => item.href !== '/#contacts')
+export const contactNavigationItems = navigationItems.filter((item) => item.href !== '/#contact')